requirements.txt
google-generativeai
python-telegram-bot
requests
packaging
rich
polib
tenacity
argparse-color-formatter